refactored code and replaced media queries with mixins
