+++ package.json
@@ -16,6 +16,7 @@
     "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
     "react-slick": "^0.29.0",
+    "sass": "^1.62.1",
     "slick-carousel": "^1.8.1",
     "web-vitals": "^2.1.4"
   },
