--- src/components/AboutSection/AboutSection.module.css
@@ -1,95 +1,0 @@
-.aboutSectionContainer {
-    padding: 80px 40px 100px 40px;
-}
-
-.contentContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: fit-content;
-    gap: 120px;
-}
-
-.content {
-    display: flex;
-    flex-direction: column;
-    max-width: 430px;
-    gap: 25px;
-}
-
-.title {
-    font-family: Georgia;
-    font-size: 35px;
-    font-weight: 400;
-    text-align: left;
-    line-height: 45.5px;
-    letter-spacing: 0.06em;
-    color: #545454;
-}
-
-.content > p {
-    font-size: 15px;
-    text-align: left;
-    line-height: 24px;
-    color: #4c4c4c;
-}
-
-.pictureContainer > img {
-    max-width: 100%;
-    height: auto;
-}
-
-@media (max-width: 768px) {
-    .aboutSectionContainer {
-        padding: 80px 165px 100px;
-    }
-    
-    .contentContainer {
-        flex-direction: column;
-        gap: 80px;
-    }
-
-    .content {
-        order: 1;
-    }
-
-    .pictureContainer {
-        order: 2;
-    }
-}
-
-@media (max-width: 470px) {
-    .aboutSectionContainer {
-        padding: 42px 25px;
-    }
-    
-    .contentContainer {
-        gap: 0;
-    }
-
-    .content {
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        gap: 0;
-    }
-
-    .title {
-        font-size: 25px;
-        line-height: 32px;
-        letter-spacing: 0.06em;
-        text-align: center;
-    }
-
-    .content p:first-of-type {
-        margin-top: 25px;
-    }
-
-    .content p:last-of-type {
-        margin-top: 22px;
-    }
-
-    .pictureContainer {
-        margin: 42px 5px 0;
-    } 
-}
