+++ src/components/AboutSection/AboutSection.module.scss
@@ -1,0 +1,95 @@
+@import '../styles/mediaMixin.scss';
+
+.aboutSectionContainer {
+    padding: 80px 40px 100px 40px;
+
+    @include tablets {
+        padding: 80px 165px 100px;
+    }
+
+    @include mobiles {
+        padding: 42px 25px;
+    }
+}
+
+.contentContainer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: fit-content;
+    gap: 120px;
+
+    @include tablets {
+        flex-direction: column;
+        gap: 80px;
+    }
+
+    @include mobiles {
+        gap: 0;
+    }
+}
+
+.title {
+    font-family: Georgia;
+    font-size: 35px;
+    font-weight: 400;
+    text-align: left;
+    line-height: 45.5px;
+    letter-spacing: 0.06em;
+    color: #545454;
+
+    @include mobiles {
+        font-size: 25px;
+        line-height: 32px;
+        letter-spacing: 0.06em;
+        text-align: center;
+    }
+}
+
+.content {
+    display: flex;
+    flex-direction: column;
+    max-width: 430px;
+    gap: 25px;
+
+    @include tablets {
+        order: 1;
+    }
+
+    @include mobiles {
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        gap: 0;
+
+        & p:first-of-type {
+            margin-top: 25px;
+        }
+
+        & p:last-of-type {
+            margin-top: 22px;
+        }
+    }
+}
+
+.content > p {
+    font-size: 15px;
+    text-align: left;
+    line-height: 24px;
+    color: #4c4c4c;
+}
+
+.pictureContainer > img {
+    max-width: 100%;
+    height: auto;
+}
+
+.pictureContainer {
+    @include tablets {
+        order: 2;
+    }
+
+    @include mobiles {
+        margin: 42px 5px 0;
+    }
+}
