--- src/components/DonationSection/DonationSection.module.css
@@ -1,138 +1,0 @@
-.donationSectionContainer {
-    height: 100%;
-    padding-top: 80px;
-    padding-bottom: 100px;
-    background: #F6F6F6;
-}
-
-.contentContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 30px;
-}
-
-.content {
-    display: flex;
-    flex-direction: column;
-    max-width: 380px;
-    gap: 20px;
-    text-align: left;
-}
-
-.header {
-    font-family: Georgia;
-    font-size: 35px;
-    font-weight: 400;
-    line-height: 45px;
-    letter-spacing: 0.06em;
-    color: #545454;
-}
-
-.subHeader {
-    font-family: Georgia;
-    font-size: 15px;
-    font-weight: 400;
-    letter-spacing: 1px;
-    color: #545454;
-}
-
-.imageContainer > img {
-    width: 100%;
-    height: auto;
-}
-
-.creditCard {
-    display: flex;
-    padding: 10px 15px;
-    width: fit-content;
-    gap: 15px;
-    background-color: #f1cdb3;
-    border-radius: 9px;
-}
-
-.creditCard > p {
-    font-family: Georgia;
-    font-size: 20px;
-    font-weight: 400;
-    font-style: normal;
-    letter-spacing: 0.06em;
-    color: #545454;
-}
-
-.legalInfo {
-    font-size: 12px;
-    font-style: italic;
-    font-weight: 400;
-    line-height: 18px;
-    color: #b2b2b2;
-}
-
-@media (max-width: 768px) {
-    .donationSectionContainer {
-        padding: 80px 60px 100px;
-    }
-
-    .contentContainer {
-        flex-direction: column;
-        gap: 60px;
-    }
-
-    .content {
-        order: 1;
-        max-width: 380px;
-    }
-
-    .imageContainer {
-        order: 2;
-    }
-}
-
-@media (max-width: 470px) {
-    .donationSectionContainer {
-        padding: 42px 10px;
-    }
-
-    .content {
-        align-items: center;
-    }
-
-    .contentContainer {
-        gap: 42px;
-    }
-    
-    .header {
-        text-align: center;
-        font-size: 25px;
-        line-height: 32px;
-    }
-
-    .subHeader {
-        text-align: center;
-        font-size: 15px;
-        line-height: 24px;
-        letter-spacing: 0;
-    }
-
-    .creditCard {
-        padding: 13px 15px;
-    }
-
-    .creditCard > h5 {
-        font-size: 15px;
-    }
-
-    .legalInfo {
-        max-width: 280px;
-        text-align: justify;
-    }
-
-    .imageContainer {
-        order: 2;
-        margin: 0 20px;
-    }
-    .imageContainer > img {
-        max-width: 100%;
-        height: auto;
-    }
-}
