+++ src/components/Header/Navbar.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import classes from './Navbar.module.css';
+import classes from './Navbar.module.scss';
 import { NavLink } from 'react-router-dom';
 
 const Navbar = ({lightStyles, menuActive}) => {
