--- src/components/Header/Navbar.module.css
@@ -1,91 +1,0 @@
-.navbarContainerDark {
-    position: relative;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    gap: 35px;
-}
-
-.navbarContainerLight {
-    position: relative;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    gap: 35px;
-}
-
-.navbarContainerDark > a {
-    color: #cdcdcd;     
-}
-
-.navbarContainerDark > a:hover {
-    color: #fafafa;
-    transition: 0.3s;  
-}
-
-.navbarContainerLight > a {
-    color: #545454;
-}
-
-.navbarContainerLight > a:hover {
-    color: #292929;
-    transition: 0.3s;
-}
-
-.link {
-    margin-top: 10px;
-    padding-bottom: 10px;
-    font-size: 15px;
-    font-weight: 400;
-    text-decoration: none;
-    border-bottom: 3px solid transparent;
-}
-
-.navbarContainerDark > .activeLink {
-    color: #fafafa;
-    transition: none;
-}
-
-.navbarContainerLight > .activeLink {
-    color: #292929;
-    transition: none;
-}
-
-.activeLink {
-    margin-top: 10px;
-    padding-bottom: 10px;
-    font-size: 15px;
-    font-weight: 400;
-    text-decoration: none;
-    border-bottom: 3px solid #f1cdb3;
-}
-
-@media (max-width: 700px) {
-    .navbarContainerDark,
-    .navbarContainerLight {
-        position: fixed;
-        z-index: 1;
-        left: -100%;
-        right: 0;
-        top: 0;
-        bottom: 0;
-        flex-direction: column;
-        justify-content: center;
-        width: 100%;
-        height: 100vh;
-        transition: all 0.5s;
-    }
-
-    .navbarContainerDark {
-        background: radial-gradient(100% 215.42% at 0% 0%, #5B483A 0%, #262425 100%);
-    }
-    
-    .navbarContainerLight {
-        background: radial-gradient(100% 215.42% at 0% 0%, #f6f6f6 0%, #fff 100%);
-    }
-
-    .menuMobileDark,
-    .menuMobileLight {
-        left: 0;
-    }
-}
