+++ src/components/Header/Navbar.module.scss
@@ -1,0 +1,89 @@
+@import '../styles/mediaMixin.scss';
+
+.navbarContainerDark,
+.navbarContainerLight {
+    position: relative;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    gap: 35px;
+
+    @include mobiles {
+        position: fixed;
+        z-index: 1;
+        left: -100%;
+        right: 0;
+        top: 0;
+        bottom: 0;
+        flex-direction: column;
+        justify-content: center;
+        width: 100%;
+        height: 100vh;
+        transition: all 0.5s;
+    }
+}
+
+.navbarContainerDark {
+    & > .activeLink {
+        color: #fafafa;
+        transition: none;
+    }
+    
+    & > a {
+        color: #cdcdcd;
+
+        &:hover {
+            color: #fafafa;
+            transition: 0.3s; 
+        }
+    }
+
+    @include mobiles {
+        background: radial-gradient(100% 215.42% at 0% 0%, #5B483A 0%, #262425 100%);
+    }
+}
+
+.navbarContainerLight {
+    & > .activeLink {
+        color: #292929;
+        transition: none;
+    }
+
+    & > a {
+        color: #545454;
+
+        &:hover {
+            color: #292929;
+            transition: 0.3s;
+        }
+    }
+
+    @include mobiles {
+        background: radial-gradient(100% 215.42% at 0% 0%, #f6f6f6 0%, #fff 100%);
+    }
+}
+
+.link {
+    margin-top: 10px;
+    padding-bottom: 10px;
+    font-size: 15px;
+    font-weight: 400;
+    text-decoration: none;
+    border-bottom: 3px solid transparent;
+}
+
+.activeLink {
+    margin-top: 10px;
+    padding-bottom: 10px;
+    font-size: 15px;
+    font-weight: 400;
+    text-decoration: none;
+    border-bottom: 3px solid #f1cdb3;
+}
+
+.menuMobileDark,
+.menuMobileLight {
+    @include mobiles {
+        left: 0;
+    }
+}
