+++ src/components/HelpSection/HelpSection.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import classes from './HelpSection.module.css';
+import classes from './HelpSection.module.scss';
 import helpList from '../../constants/helpList';
 
 const HelpSection = () => {
     return (
         <div className={classes.helpSectionContainer}>
-            <h3 className={classes.title}>How you can help<br/>our shelter</h3>
+            <h3 className={classes.title}>How you can help our shelter</h3>
 
             <div className={classes.helpList}>
                 {
