+++ src/components/HelpSection/HelpSection.module.scss
@@ -1,7 +1,23 @@
+@import '../styles/mediaMixin.scss';
+
 .helpSectionContainer {
     position: relative;
     height: 100%;
     padding: 80px 40px 100px;
+
+    @include tablets {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        flex-direction: column;
+        gap: 60px;
+        padding: 80px 69px 100px;
+    }
+
+    @include mobiles {
+        padding: 42px 15px;
+        gap: 42px;
+    }
 }
 
 .title {
@@ -11,6 +27,11 @@
     line-height: 45px;
     letter-spacing: 0.06em;
     color: #545454;
+
+    @include mobiles {
+        max-width: 220px;
+        font-size: 25px;
+    }
 }
 
 .helpItem {
@@ -18,6 +39,26 @@
     flex-direction: column;
     align-items: center;
     gap: 30px;
+
+    & > p {
+        font-family: Georgia;
+        font-size: 20px;
+        font-weight: 400;
+        letter-spacing: 1px;
+        color: #545454;
+    }
+
+    @include tablets {
+        width: 170px;
+    }
+
+    @include mobiles {
+        width: 130px;
+
+        & > p {
+            font-size: 15px;
+        }
+    }
 }
 
 .iconContainer {
@@ -28,14 +69,6 @@
     height: 60px;
 }
 
-.helpItem > p {
-    font-family: Georgia;
-    font-size: 20px;
-    font-weight: 400;
-    letter-spacing: 1px;
-    color: #545454;
-}
-
 .helpList {
     display: flex;
     justify-content: center;
@@ -44,50 +77,16 @@
     padding: 60px 87px 0;
     margin: 0 auto;
     max-width: 1026px;
-}
 
-/* 1052px is used instead of 768px */
-@media (max-width: 1052px) {
-    .helpSectionContainer {
-        display: flex;
-        flex-direction: column;
-        gap: 60px;
-        padding: 80px 69px 100px;
-    }
-
-    .helpItem {
-        width: 170px;
-    }
-    
-    .helpList {
+    @include tablets {
         justify-content: center;
         gap: 55px 60px;
         margin: 0;
         padding: 0;
     }
-}
 
-@media (max-width: 470px) {
-    .helpSectionContainer {
-        padding: 42px 15px;
-        gap: 42px;
-    }
-
-    .helpList {
-        gap: 30px 30px;
-    }
-
-    .helpItem {
-        max-width: 130px;
-    }
-
-    .helpItem > p {
-        font-size: 15px;
-    }
-}
-
-@media (max-width: 319px) {
-    .helpList {
+    @include mobiles {
         justify-content: center;
+        gap: 30px 30px;
     }
 }
