+++ src/components/Pagination/PaginationSection.module.scss
@@ -1,0 +1,117 @@
+@import '../styles/mediaMixin.scss';
+
+.carouselContainer {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 60px;
+    padding: 80px 40px 100px;
+    background-color: #f6f6f6;
+
+    @include tablets {
+        padding: 60px 30px 70px;  
+        gap: 30px; 
+    }
+
+    @include mobiles {
+        padding: 42px 10px;
+        gap: 42px;
+    }
+}
+
+.title {
+    font-family: Georgia;
+    font-size: 35px;
+    font-weight: 400;
+    font-style: normal;
+    line-height: 45px;
+    letter-spacing: 0.06em;
+    color: #545454;
+
+    @include mobiles {
+        font-size: 25px;
+        line-height: 32px;
+    }
+}
+
+.itemsContainer {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    gap: 30px 40px;
+    max-width: 1200px;
+
+    @include tablets {
+        gap: 30px 40px;
+    }
+
+    @include mobiles {
+        flex-direction: column;
+        align-items: center;
+        gap: 30px;
+    }
+}
+
+.paginationButtonsWrapper,
+.paginationWrapper {
+    display: flex;
+    align-items: center;
+    gap: 20px;
+
+    @include mobiles {
+        gap: 10px;
+    }
+}
+
+.paginationButtonsWrapper > li {
+    display: flex;
+}
+
+.disabledButton,
+.pageContainer,
+.paginationButton {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 50px;
+    height: 50px;
+    border: 1px solid #f1cdb3;
+    border-radius: 100px;
+    background: transparent;
+}
+
+.paginationWrapper > .paginationButton {
+    width: 52px;
+    height: 52px;
+    text-align: center;
+}
+
+.pageContainer {
+    border: 1px solid #f1cdb3;
+    background-color: #f1cdb3;
+}
+
+.disabledButton {
+    width: 52px;
+    height: 52px;
+    border: 1px solid #cdcdcd;
+    border-color: #cdcdcd;
+    background-color: transparent;
+}
+
+.disabledLink > a {
+    border: 1px solid #cdcdcd;
+    background-color: transparent;
+}
+
+.disabledLink > a:hover {
+    border: 1px solid #cdcdcd;
+    background-color: transparent;
+}
+
+.paginationButton:hover {
+    border-color: #fddcc4;
+    background-color: #fddcc4;
+    transition: 0.3s;
+}
