+++ src/components/PetsSection/Carousel.module.scss
@@ -1,12 +1,24 @@
+@import '../styles/mediaMixin.scss';
+
 .sliderContainer {
     display: flex;
     align-items: center;
     justify-content: space-between;
     height: fit-content;
+
+    @include mobiles {
+        flex-wrap: wrap;
+        justify-content: center;
+        gap: 20px 80px;
+    }
 }
 
 .slickSlider {
     width: 90%;
+
+    @include mobiles {
+        width: 100%;
+    }
 }
 
 .listItemWrapper {
@@ -29,19 +41,9 @@
     overflow: hidden;
 }
 
-@media (max-width: 470px) {
-    .sliderContainer {
-        flex-wrap: wrap;
-        justify-content: center;
-        gap: 20px 80px;
-    }
-
-    .slickSlider {
-        width: 100%;
-    }
-
+@include mobiles {
     .prevButton {
-        order: 2; 
+        order: 2;
     }
 
     .nextButton {
