+++ src/components/PetsSection/PetsSection.module.scss
@@ -1,3 +1,5 @@
+@import '../styles/mediaMixin.scss';
+
 .petsSection {
     display: flex;
     flex-direction: column;
@@ -7,15 +9,32 @@
     padding: 80px 40px 100px 40px;
     background-color: #f6f6f6;
     animation: appear 3s;
+
+    @include tablets {
+        padding: 80px 30px 100px;
+    }
+
+    @include mobiles {
+        padding: 42px 10px;
+        gap: 42px;
+    }
 }
 
-.header > h3 {
+.title {
+    max-width: 400px;
     font-family: Georgia;
     font-size: 35px;
     font-weight: 400;
+    text-align: center;
     line-height: 45px;
     letter-spacing: 0.06em;
     color: #545454;
+
+    @include mobiles {
+        font-size: 25px;
+        line-height: 32px;
+        text-align: center;
+    }
 }
 
 .sliderSection {
@@ -23,6 +42,10 @@
     align-items: center;
     justify-content: center;
     width: 100%;
+
+    @include mobiles {
+        margin: 0 15px;
+    }
 }
 
 .footerButton {
@@ -41,27 +64,4 @@
 .footerButton:hover {
     background-color: #FDDCC4;
     transition: 0.3s;
-}
-
-@media (max-width: 768px) {
-    .petsSection {
-        padding: 80px 30px 100px;
-    }
-}
-
-@media (max-width: 470px) {
-    .petsSection {
-        padding: 42px 10px;
-        gap: 42px;
-    }
-
-    .sliderSection {
-        margin: 0 15px;
-    }
-    
-    .header > h3 {
-        font-size: 25px;
-        line-height: 32px;
-        text-align: center;
-    }
 }
