+++ src/components/StartSection/StartSection.module.scss
@@ -1,3 +1,5 @@
+@import '../styles/mediaMixin.scss';
+
 .startScreenContainer {
     width: 100%;
     height: 100%;
@@ -16,11 +18,33 @@
     margin: 0 auto;
     padding: 30px 40px 0;
     animation: fadeIn 1s;
+
+    @include tablets {
+        position: relative;
+        justify-content: center;
+        flex-direction: column;
+        gap: 100px;
+        margin: 0;
+        padding: 30px 100px 0;
+    }
+
+    @include burgerMenu {
+        gap: 100px;
+        padding: 30px 100px 0;
+    }
 }
 
 .contentContainer > img {
     width: 100%;
     height: auto;
+
+    @include tablets {
+        align-self: center;
+        max-width: 569px;
+        height: auto;
+        margin: 0;
+        margin-left: 70px;
+    }
 }
 
 .content {
@@ -58,6 +82,10 @@
     border-width: 0;
     border-radius: 100px;
     background-color: #f1cdb3;
+
+    @include tablets {
+        align-self: center;
+    }
 }
 
 .content > button:hover {
@@ -71,39 +99,6 @@
     }
     100% {
         opacity: 1;
-    }
-}
-
-@media (max-width: 768px) {
-    .contentContainer {
-        position: relative;
-        justify-content: center;
-        flex-direction: column;
-        gap: 100px;
-        margin: 0;
-        padding: 30px 100px 0;
-    }
-
-    .content > button {
-        align-self: center;
-    }
-
-    .contentContainer {
-        position: relative;
-    }
-
-    .contentContainer > img {
-        max-width: 569px;
-        height: auto;
-        margin: 0;
-        margin-left: 70px;
-    }
-}
-
-@media (max-width: 700px) {
-    .contentContainer {
-        gap: 100px;
-        padding: 30px 100px 0;
     }
 }
