+++ src/components/styles/mediaMixin.scss
@@ -1,0 +1,23 @@
+@mixin tablets {
+    @media (max-width: 1280px) {
+        @content;
+    }
+}
+
+@mixin burgerMenu {
+    @media (max-width: 700px) {
+        @content;
+    }
+}
+
+@mixin biggerMobiles {
+    @media (max-width: 767px) {
+        @content;
+    }
+}
+
+@mixin mobiles {
+    @media (max-width: 767px) {
+        @content;
+    }
+}
